Namespace function seeded_rand_string()
